fix(#1272): deduplicate openShell/declinedIndices fillet fixtures
- Create FilletTestFixtures.swift with shared openShell() and declinedIndices
- Update Issue612FilletContourSelectionTests, Issue633BlendedEdgesDuplicateReportTests, Issue639FilletDeclinedEdgeReportTests to use shared fixtures
- All 23 fillet tests pass
- Gate scripts pass
